Revert "Merge branch 'develop' into feature/22_top_page"
This reverts commit 81f63a6bf52c57d7290ea6768c18e91682c8684f, reversing
changes made to b1f2713b298ff504d8beac3214141e38ec7a7666.

+++ app/controllers/api/v1/users_controller.rb
@@ -2,18 +2,17 @@ class Api::V1::UsersController < Api::V1::BaseController
   before_action :require_login, only: %i[update destroy]
 
   def index
-    @pagy, users = pagy(User.published)
+    @pagy, users = pagy(User.all)
     render json: users
   end
 
   def show
-    user = User.find_by!(uuid: params[:uuid])
-    authorize!(user)
+    user = User.find_by(uuid: params[:uuid])
     render json: user
   end
 
   def update
-    user = User.find_by!(uuid: params[:uuid])
+    user = User.find_by(uuid: params[:uuid])
     return head 404 unless user == current_user
 
     user.assign_attributes(user_params)
@@ -22,19 +21,27 @@ class Api::V1::UsersController < Api::V1::BaseController
       render json: user
     else
       error_json = {
-        'code' => '422',
+        'status' => '422',
         'title' => '登録内容が適切ではありません',
         'detail' => '登録内容を確認してください',
         'messages' => user.errors.full_messages
       }
-      render json: { 'error': error_json }, status: :unprocessable_entity
+      render json: { 'error': error_json }, status: 422
     end
   end
 
   def destroy
-    user = User.find_by!(uuid: params[:uuid])
-    authorize!(user)
-    user.destroy!
+    user = User.find_by(uuid: params[:uuid])
+    if user == current_user
+      user.destroy!
+    else
+      head 404
+    end
+  end
+
+  def current
+    user = current_user if logged_in?
+    render json: user
   end
 
   private
